pyproject.toml
[tool.poetry]
name = "django-graphql-benchmarks"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
requests = "^2.26.0"
Django = "^3.2.6"
strawberry-graphql = "^0.72.3"
graphene = "3.0.0b7"
djangorestframework = "^3.12.4"
graphene-django = "3.0.0b7"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
numpy = "^1.21.2"
dash = "^1.21.0"
dash-core-components = "^1.17.1"
dash-html-components = "^1.1.4"
dash-renderer = "^1.9.1"
dash-table = "^4.12.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.4.3"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
